# Reduce leakage and improve performance of explosion intersections.