--- a/include/intersectables/explosion.h
+++ b/include/intersectables/explosion.h
@@ -17,10 +17,10 @@ class Explosion : public Intersectable {
     Explosion(Material*, const Point3f, float);
 
     virtual HitRecord* intersect(Ray* ray) const;
-    float signed_distance(Point3f* p) const;
-    float fractal_brownian_motion(Point3f* x) const;
-    Point3f rotate(Point3f* v) const;
-    float noise(Point3f* x) const;
+    float signed_distance(const Point3f& p) const;
+    float fractal_brownian_motion(const Point3f& x) const;
+    Point3f rotate(const Point3f& v) const;
+    float noise(const Point3f& x) const;
 
     template <typename T> inline T lerp(const T &v0, const T &v1, float t) const {
       return v0 + (v1-v0)*std::max(0.f, std::min(1.f, t));