--- a/src/intersectables/explosion.cpp
+++ b/src/intersectables/explosion.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 #include "explosion.h"
 
-Explosion::Explosion(Material* material, Point3f* center, float radius)
+Explosion::Explosion(Material* material, const Point3f center, float radius)
   : material(material), center(center), radius(radius)
 {}