Again by making more use of stack variables. Allocating to heap is expensive. Timings:

Before:
BM_ExplosionIntersect    1742571 ns      1741476 ns          400

After:
BM_ExplosionIntersect     284027 ns       283994 ns         2452